Better Lighting
Swapped "point light" method of lighting the solar system with a directional light that rotates to look in the direction of the player. Accurately simulating light over large distances is too complex and this is a good enough work around for visual purposes. This also allows planets to actually cast shadows on rings

NOTE: If the player gets too close to a celestial, they will cause a 'shadow' which blocks the light. The current fix is to turn up the intensity of the spotlight on the player until can figure out how to stop player casting shadows (this can be seen in the changes of the controller PFs which didnt work)